tests/wpt/tests/html/semantics/scripting-1/the-script-element/async_001.htm

<!DOCTYPE html>
<html>
    <head>
        <title>Async property on a dynamically-created script is true by default</title>
        <meta http-equiv="content-type" content="text/html; charset=UTF-8" />
        <meta description="This test checks the Async property on a dynamically-created script element. By default it should be true." />
        <link rel="author" title="Microsoft" href="http://www.microsoft.com/" />
        <link rel="help" href="https://html.spec.whatwg.org/multipage/#dom-script-async"/>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
    </head>
    <body>
        <div id=log></div>
        <script type="text/javascript">
            test(function() {assert_true(document.createElement("script").async)}, "Async property on a dynamically-created script is true by default");
        </script>
    </body>
</html>
